Extract keys from common
